requirements.txt
datasets>=2.20.0
huggingface_hub>=0.24.0
pandas>=2.2.0
pyarrow>=16.0.0
numpy>=1.26.0
scipy>=1.12.0
scikit-learn>=1.4.0
tqdm>=4.66.0
matplotlib>=3.8.0
pyyaml>=6.0.1
